collections/ansible_collections/beyondtrust/secrets_safe/requirements.txt
ansible-core>=2.14
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.3
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
pycparser==2.21
pyOpenSSL==23.2.0
pytest==7.4.0
PyYAML==6.0
requests==2.31.0
resolvelib==1.0.1
retry-requests==2.0.0
urllib3==2.0.3